backend/requirements.txt
# Dubai AI Lost & Found - Backend Requirements

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin to version compatible with passlib
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AI / ML
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchvision==0.17.0+cpu
pillow==10.2.0
numpy<2.0  # PyTorch 2.2.0 requires NumPy 1.x

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Optional: For future privacy features
# opencv-python==4.9.0.80

# AI/ML Utilities
ultralytics
easyocr
